Move verticalFormater to filter.go
In previous commit, the type of verticalFormater was changed to Filter.
So move it to filter.go.
